rename strategy -> executionStrategy
