+++ src/main/executionStrategy/index.ts
@@ -1,0 +1,16 @@
+import type { Demo } from '../../common/protocol'
+
+export type ChangeDemo = (demo: Demo) => void
+export type ExecutionStrategyDestroy = () => void
+export interface ExecutionStrategy {
+  destroy: ExecutionStrategyDestroy
+  changeDemo: ChangeDemo
+}
+
+export type SetFatalError = (message: string) => void
+export interface ExecutionStrategyStartOptions {
+  setFatalError: SetFatalError
+  canvasElement: HTMLCanvasElement
+  initialDemo: Demo
+}
+export type ExecutionStrategyStart = (options: ExecutionStrategyStartOptions) => ExecutionStrategy
